+++ .gitignore
@@ -4,7 +4,7 @@ config.yml
 /config/
 /cache/
 database.db.bkp
-database.db
+db.sqlite3
 logs/
 data.json
 external_modules/
