+++ bot/__init__.py
@@ -2,12 +2,10 @@
 # -*- coding: utf-8 -*-
 
 from .events import CommandEvent, MessageEvent, BotMentionEvent
-from .libs.configuration import StaticConfig, Configuration, BaseModel
+from .libs.dynamic_config import Configuration, BaseModel
+from .libs.configuration import StaticConfig
 from .libs.language import Language, SingleLanguage
 from .command import Command
 from .manager import Manager
-
-from .libs.configuration import get_database, init_db
-from . import defaults
 
 from .bot import AlexisBot
