+++ bot/libs/configuration.py
@@ -1,13 +1,10 @@
 import os
-import discord
-import peewee
-from playhouse.db_url import connect
+
 from ruamel.yaml import YAML
 
+from bot import defaults
 
-db = None
 yaml = YAML(typ='safe')
-yaml.default_flow_style = False
 
 
 class StaticConfig:
@@ -35,10 +32,10 @@ class StaticConfig:
     def __contains__(self, item):
         return item in self.config
 
-    def load(self, defaults=None):
+    def load(self, default_config=None):
         """
         Loads the configuration file
-        :param defaults: Default settings to load into the file
+        :param default_config: Default settings to load into the file
         :return: A dict with the file data and the default settings, if passed.
         """
         with open(self.path) as f:
@@ -48,19 +45,19 @@ class StaticConfig:
 
         self.config = loaded_conf
 
-        if defaults is not None and isinstance(defaults, dict):
-            self.config = {**defaults, **loaded_conf}
+        if default_config is not None and isinstance(default_config, dict):
+            self.config = {**default_config, **loaded_conf}
             self.save()
 
-    def load_defaults(self, defaults):
+    def load_defaults(self, default_config):
         """
         Loads default settings into the file
-        :param defaults: A dict with the defaults
+        :param default_config: A dict with the defaults
         """
-        if not isinstance(defaults, dict):
+        if not isinstance(default_config, dict):
             raise RuntimeError('defaults argument must be a dict instance')
 
-        self.config = {**defaults, **self.config}
+        self.config = {**default_config, **self.config}
         self.save()
 
     def save(self, reload=False):
@@ -126,326 +123,17 @@ class StaticConfig:
         return os.path.exists('config/' + name + '.yml')
 
 
-class Configuration:
-    """
-    Manages per-server configuration on the database, cached on memory.
-    """
+class BotConfiguration(StaticConfig):
     def __init__(self):
-        """
-        Initialize and fetch all the configuration from the database, then store it on memory.
-        """
-        self.q = ServerConfig.select().execute()
-        self.sv = {}
+        super().__init__(path='config.yml')
+        self.load(defaults.config)
+        self.loaded = True
 
-        for conf in list(self.q):
-            if conf.serverid not in self.sv:
-                self.sv[conf.serverid] = {}
-            self.sv[conf.serverid][conf.name] = conf.value
+    _instance = None
 
-    def has(self, svid, name):
-        """
-        Checks if the server has a configuration value, without setting a default value.
-        :param svid: The server ID
-        :param name: The configuration value name
-        :return: A boolean value from the operation result.
-        """
-        try:
-            if svid not in self.sv:
-                self.sv[svid] = {}
+    @staticmethod
+    def get_instance():
+        if BotConfiguration._instance is None:
+            BotConfiguration._instance = BotConfiguration()
 
-            if name in self.sv[svid]:
-                return True
-
-            ServerConfig.get(serverid=svid, name=name)
-            return True
-        except ServerConfig.DoesNotExist:
-            return False
-
-    def get(self, svid, name, default='', create=True):
-        """
-        Fetch a configuration value for a server. If the configuration does not exist, it will be initialized
-        with the default value, then, is stored on memory. If the default value is None and the value does not
-        exist, then the database will not be queried.
-        :param svid: The server ID
-        :param name: The value name to retrieve
-        :param default: The default value to use if the configuration does not exist
-        :param create: If the value does not exist, and the default value is passed, the value will be created
-        on the database and cached on memory. Else, the default value is returned and nothing else.
-        :return: The requested configuration value.
-        """
-        if svid not in self.sv:
-            self.sv[svid] = {}
-
-        if name not in self.sv[svid]:
-            if default is None and not self.has(svid, name):
-                return None
-
-            if create:
-                q, _ = ServerConfig.get_or_create(serverid=svid, name=name, defaults={'value': str(default)})
-                self.sv[svid][name] = q.value
-            else:
-                return default or None
-
-        return self.sv[svid][name]
-
-    def set(self, svid, name, value):
-        """
-        Stores a configuration in database and memory. If the value is the same as in memory, then the database is
-        not altered.
-        :param svid: The server ID
-        :param name: The configuration value to be set
-        :param value: The new configuration value
-        :return: The stored value
-        """
-        value = str(value)
-        if self.get(svid, name, value) == value:
-            return value
-
-        q, _ = ServerConfig.get_or_create(serverid=svid, name=name)
-        q.value = value
-        q.save()
-        self._local_save(svid, name, value)
-
-        return q.value
-
-    def unset(self, svid, name):
-        """
-        Deletes a configuration from the database. It does not raise any exceptions.
-        :param svid: The server ID
-        :param name: The configuration value name
-        :return: A boolean given if the value existed or not
-        """
-        if self.has(svid, name):
-            try:
-                ins = ServerConfig.get(serverid=svid, name=name)
-                ins.delete_instance()
-                del self.sv[svid][name]
-                return True
-            except ServerConfig.DoesNotExist:
-                return False
-        return False
-
-    def get_list(self, svid, name, separator=',', default=None):
-        """
-        Fetches a configuration value as a list
-        :param svid: The server ID
-        :param name: The configuration value name
-        :param separator: The separator that will split the values (as it's stored as a single string)
-        :param default: The default value to return if the config does not exist.
-        :return: The requested configuration as a list
-        """
-        if default is None:
-            default = []
-        val = self.get(svid, name, '')
-        return default if val == '' else val.split(separator)
-
-    def set_list(self, svid, name, elements, separator=','):
-        """
-        Stores a list in the configuration, overwriting current values.
-        :param svid: The server ID
-        :param name: The configuration value name
-        :param elements: The list to be stored
-        :param separator: The glue for the items list (as it's stored as a single string)
-        """
-        if not isinstance(elements, list):
-            raise RuntimeError('elements argument only supports a list instance')
-
-        result = separator.join(elements)
-        self.set(svid, name, result)
-
-    def add(self, svid, name, value, separator=','):
-        """
-        Fetches a cofiguration value as a list, appends a value, then stores it.
-        :param svid: The server ID
-        :param name: The configuration value name
-        :param value: The value to be added. If it's already on the list, it won't be added.
-        :param separator: The splitter string
-        :return: The list with the new value added (if it wasn't on the list already)
-        """
-        values = self.get_list(svid, name, separator)
-        if value not in values:
-            values.append(value)
-            self.set_list(svid, name, values, separator)
-
-        return values
-
-    def remove(self, svid, name, value, separator=','):
-        """
-        Fetches a cofiguration value as a list, removes a value if it exists, then stores the list.
-        :param svid: The server ID
-        :param name: The configuration value name
-        :param value: The value to be removed from the list
-        :param separator: The splitter string (as the value is stored as a single string)
-        :return: The list with the value removed, or the same list if the item wasn't on the list.
-        """
-        values = self.get_list(svid, name, separator)
-        if value in values:
-            values.remove(value)
-            self.set_list(svid, name, values, separator)
-
-        return values
-
-    def remove_index(self, svid, name, idx, separator):
-        """
-        Fetches a cofiguration value as a list, removes a value by it's index, then stores the list.
-        :param svid: The server ID
-        :param name: The configuration value name
-        :param idx: The value index, starting from zero. If the index is invalid, the list won't be modified.
-        :param separator: The splitter string (as the value is stored as a single string)
-        :return: The list with the value removed, or the same list if the item wasn't on the list.
-        """
-        values = self.get_list(svid, name, separator)
-        if abs(values) >= len(values):
-            return values
-
-        del values[idx]
-        self.set_list(svid, name, values, separator)
-        return values
-
-    def _local_save(self, svid, name, value):
-        """
-        Guarda una configuración en memoria solamente
-        :param svid: El ID del servidor
-        :param name: El nombre de la configuración
-        :param value: El valor de la configuración
-        :return: El valor de la configuración guardada
-        """
-        if svid not in self.sv:
-            self.sv[svid] = {}
-
-        self.sv[svid][name] = str(value)
-        return self.sv[svid][name]
-
-
-class GuildConfiguration:
-    """
-    Shortcut to manage a single server configuration from a Configuration instance.
-    """
-    def __init__(self, mgr, guild: discord.Guild = None):
-        """
-        :param mgr: The Configuration instance for the global configuration manager
-        :param guild: The discord.Server instance or server ID
-        """
-        self.guild_id = guild.id if guild is not None else 'all'
-        self.mgr = mgr
-
-    def has(self, name):
-        """
-        Checks if the server has a configuration value, without setting a default value.
-        :param name: The configuration value name
-        :return: A boolean value from the operation result.
-        """
-        return self.mgr.has(self.guild_id, name)
-
-    def get(self, name, default='', create=True):
-        """
-        Fetch a configuration value for a server. If the configuration does not exist, it will be initialized
-        with the default value, then, is stored on memory. If the default value is None and the value does not
-        exist, then the database will not be queried.
-        :param name: The value name to retrieve
-        :param default: The default value to use if the configuration does not exist
-        :param create: If the value does not exist, and the default value is passed, the value will be created
-        on the database and cached on memory. Else, the default value is returned and nothing else.
-        :return: The requested configuration value.
-        """
-        return self.mgr.get(self.guild_id, name, default, create)
-
-    def set(self, name, value):
-        """
-        Stores a configuration in database and memory. If the value is the same as in memory, then the database is
-        not altered.
-        :param name: The configuration value to be set
-        :param value: The new configuration value
-        :return: The stored value
-        """
-        return self.mgr.set(self.guild_id, name, value)
-
-    def unset(self, name):
-        """
-        Deletes a configuration value from the database. It does not raise any exceptions.
-        :param name: The configuration value name
-        :return: A boolean given if the value existed or not
-        """
-        return self.mgr.unset(self.guild_id, name)
-
-    def get_list(self, name, separator=',', default=None):
-        """
-        Fetches a configuration value as a list
-        :param name: The configuration value name
-        :param separator: The separator that will split the values (as it's stored as a single string)
-        :param default: The default value if it does not exist.
-        :return: The requested configuration as a list
-        """
-        return self.mgr.get_list(self.guild_id, name, separator, default)
-
-    def set_list(self, name, elements, separator='.'):
-        """
-        Stores a list in the configuration, overwriting current values.
-        :param name: The configuration value name
-        :param elements: The list to be stored
-        :param separator: The glue for the items list (as it's stored as a single string)
-        """
-        self.mgr.set_list(self.guild_id, name, elements, separator)
-
-    def add(self, name, value, separator=','):
-        """
-        Fetches a cofiguration value as a list, appends a value, then stores it.
-        :param name: The configuration value name
-        :param value: The value to be added. If it's already on the list, it won't be added.
-        :param separator: The splitter string
-        :return: The list with the new value added (if it wasn't on the list already)
-        """
-        return self.mgr.add(self.guild_id, name, value, separator)
-
-    def remove(self, name, value, separator=','):
-        """
-        Fetches a cofiguration value as a list, removes a value if it exists, then stores the list.
-        :param name: The configuration value name
-        :param value: The value to be removed from the list
-        :param separator: The splitter string (as the value is stored as a single string)
-        :return: The list with the value removed, or the same list if the item wasn't on the list.
-        """
-        return self.mgr.remove(self.guild_id, name, value, separator)
-
-    def remove_index(self, name, idx, separator=','):
-        """
-        Fetches a cofiguration value as a list, removes a value by it's index, then stores the list.
-        :param name: The configuration value name
-        :param idx: The value index, starting from zero. If the index is invalid, the list won't be modified.
-        :param separator: The splitter string (as the value is stored as a single string)
-        :return: The list with the value removed, or the same list if the item wasn't on the list.
-        """
-        return self.mgr.remove(self.guild_id, name, idx, separator)
-
-
-def get_database():
-    global db
-    cfg = StaticConfig('config.yml')
-    cfg.load({'database_url': 'sqlite:///database.db'})
-
-    dburl = cfg['database_url']
-    if dburl.startswith('mysql:'):
-        dburl += '&amp;' if '?' in dburl else '?'
-        dburl += 'charset=utf8mb4;'
-
-    if db is None:
-        db = connect(dburl)
-
-    return db
-
-
-class BaseModel(peewee.Model):
-    class Meta:
-        database = get_database()
-
-
-class ServerConfig(BaseModel):
-    serverid = peewee.TextField()
-    name = peewee.TextField()
-    value = peewee.TextField(default='')
-
-
-def init_db():
-    tdb = get_database()
-    tdb.create_tables([ServerConfig], safe=True)
-    return tdb
+        return BotConfiguration._instance
