+++ bot/utils.py
@@ -5,17 +5,7 @@ import discord
 import re
 from discord import Embed, Colour
 
-pat_tag = re.compile(r'^<(@!?|#|a?:([a-zA-Z0-9\-_]+):)(\d{10,19})>$')
-pat_usertag = re.compile(r'^<@!?(\d{10,19})>$')
-pat_channel = re.compile(r'^<#(\d{10,19})>$')
-pat_subreddit = re.compile(r'^[a-zA-Z0-9_\-]{2,25}$')
-pat_emoji = re.compile(r'<a?(:([a-zA-Z0-9\-_]+):)([0-9]+)>')
-pat_normal_emoji = re.compile(r'^:[a-zA-Z\-_]+:$')
-pat_snowflake = re.compile(r'^\d{10,19}$')
-pat_colour = re.compile(r'^#?[0-9a-fA-F]{6}$')
-pat_delta = re.compile(r'^([0-9]+[smhd])+$')
-pat_delta_each = re.compile(r'([0-9]+[smhd])+')
-pat_invite = re.compile(r'(?:https?://)?(discord(?:app\.com/invite|\.gg|\.me)/[a-zA-Z0-9]+)')
+from bot.regex import pat_tag, pat_usertag, pat_channel, pat_emoji, pat_colour, pat_delta_each, pat_invite
 
 colour_list = ['default', 'teal', 'dark_teal', 'green', 'dark_green', 'blue', 'dark_blue', 'purple',
                'dark_purple', 'gold', 'dark_gold', 'orange', 'dark_orange', 'red', 'dark_red',
