+++ lang.py
@@ -2,8 +2,7 @@ import re
 import sys
 from os import path, listdir, mkdir
 
-from bot import Manager
-from bot.libs.configuration import StaticConfig
+from bot import Manager, StaticConfig
 from bot.utils import is_int
 
 LANG_FOLDER = path.join(Manager.get_bot_root(), 'lang')
