+++ modules/macros.py
@@ -4,9 +4,8 @@ from datetime import datetime
 import peewee
 from discord import Embed, Colour
 
-from bot import Command, categories
+from bot import Command, categories, BaseModel
 from bot.utils import is_int, get_colour, format_date, colour_list
-from bot.libs.configuration import BaseModel
 
 
 class MacroSet(Command):
